feat: 상품 id로 상품 수정 기능 구현
- 다음번에 구현할것
1) id 값이 없으면 없는 상품이라고 뜨게 하기

+++ src/main/java/gift/dto/EditProduct.java
@@ -6,19 +6,19 @@ public class EditProduct {
 
     public static class Request{
 
-        private  long id;
+        private  int id;
         private String name;
         private int price;
         private String imageUrl;
 
-        public Request( long id , String name, int price,String imageUrl) {
+        public Request( int id , String name, int price,String imageUrl) {
             this.id = id;
             this.name=name;
             this.price=price;
             this.imageUrl=imageUrl;
         }
 
-        public long getId() {return id;}
+        public int getId() {return id;}
         public String getName() {
             return name;
         }
